backend/requirements.txt
asgiref==3.9.1
Django==5.2.4
djangorestframework==3.16.0
django-filter==24.3
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.0
sqlparse==0.5.3
